pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-mcp"
version = "0.1.0"
description = "RAG (Retrieval-Augmented Generation) + MCP (Model Context Protocol) framework"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "RAG MCP Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core Framework
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-azure-ai>=0.1.0",
    "langgraph>=0.2.0",

    # MCP Protocol
    "mcp>=0.9.0",
    "jiter>=0.5.0",

    # Vector Stores
    "chromadb>=0.5.0",
    "qdrant-client>=1.12.0",
    "pinecone-client>=5.0.0",

    # LLM Providers
    "openai>=1.50.0",
    "azure-ai-inference>=1.0.0",
    "anthropic>=0.40.0",
    "zhipuai>=2.0.0",

    # Embedding
    "sentence-transformers>=3.0.0",
    "FlagEmbedding>=1.2.0",

    # Reranking
    "colbert-ai>=0.2.0",
    "cross-engine>=0.1.0",

    # PDF Processing
    "markitdown>=0.0.1a2",

    # Data Processing
    "pypdf>=5.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",

    # Observability
    "rich>=13.0.0",
    "uvicorn>=0.30.0",
    "streamlit>=1.35.0",
    "pandas>=2.0.0",
    "plotly>=5.0.0",

    # Evaluation
    "ragas>=0.1.0",
    "deepeval>=0.7.0",

    # Utilities
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-check>=2.4.0",
    "pytest-tempdir>=0.7.0",
    "pytest-bdd>=7.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
rag-ingest = "scripts.ingest:main"
rag-query = "scripts.query:main"
rag-evaluate = "scripts.evaluate:main"
rag-dashboard = "scripts.start_dashboard:main"
rag-mcp-server = "main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]

[tool.ruff]
target-version = "py312"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "W191",   # indentation contains tabs
]
unfixable = ["B"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG", "S101"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "qdrant_client.*",
    "pinecone.*",
    "langchain.*",
    "langgraph.*",
    "markitdown.*",
    "ragas.*",
    "deepeval.*",
]
ignore_missing_imports = true
